Extract common components to separate file
Extracted `APY` and `ShareOfPool` to a separate components to use these
in Keep only pool UI.
