+++ solidity/dashboard/src/components/LiquidityRewardCard.jsx
@@ -1,6 +1,5 @@
-import React, { useMemo, useCallback } from "react"
+import React, { useMemo } from "react"
 import CountUp from "react-countup"
-import BigNumber from "bignumber.js"
 import DoubleIcon from "./DoubleIcon"
 import * as Icons from "./Icons"
 import { SubmitButton } from "./Button"
@@ -10,8 +9,8 @@ import Tooltip from "./Tooltip"
 import Banner from "./Banner"
 import { toTokenUnit } from "../utils/token.utils"
 import { gt } from "../utils/arithmetics.utils"
-import { formatPercentage } from "../utils/general.utils"
 import { LIQUIDITY_REWARD_PAIRS } from "../constants/constants"
+import { APY, ShareOfPool } from "./liquidity"
 
 const LiquidityRewardCard = ({
   title,
@@ -35,36 +34,6 @@ const LiquidityRewardCard = ({
   isAPYFetching,
   pool,
 }) => {
-  const formattedApy = useMemo(() => {
-    const bn = new BigNumber(apy).multipliedBy(100)
-    if (bn.isEqualTo(Infinity)) {
-      return Infinity
-    } else if (bn.isLessThan(0.01) && bn.isGreaterThan(0)) {
-      return 0.01
-    } else if (bn.isGreaterThan(999)) {
-      return 999
-    }
-
-    return formatPercentage(bn)
-  }, [apy])
-
-  const formattedPercentageOfTotalPool = useMemo(() => {
-    const bn = new BigNumber(percentageOfTotalPool)
-    return bn.isLessThan(0.01) && bn.isGreaterThan(0)
-      ? 0.01
-      : formatPercentage(bn)
-  }, [percentageOfTotalPool])
-
-  const formattingFn = useCallback((value) => {
-    let prefix = ""
-    if (value === 0.01) {
-      prefix = `<`
-    } else if (value >= 999) {
-      prefix = `>`
-    }
-    return `${prefix}${value}%`
-  }, [])
-
   const hasWrappedTokens = useMemo(() => gt(wrappedTokenBalance, 0), [
     wrappedTokenBalance,
   ])
@@ -148,58 +117,20 @@ const LiquidityRewardCard = ({
         } mt-2 mb-2`}
       >
         <div className={"liquidity__info-tile bg-mint-10"}>
-          <Tooltip
-            simple
-            delay={0}
-            triggerComponent={Icons.MoreInfo}
-            className={"liquidity__info-tile__tooltip"}
-          >
-            Pool APY is calculated using the&nbsp;
-            <a
-              target="_blank"
-              rel="noopener noreferrer"
-              href={"https://thegraph.com/explorer/subgraph/uniswap/uniswap-v2"}
-              className="text-white text-link"
-            >
-              Uniswap subgraph API
-            </a>
-            &nbsp;to fetch the total pool value and KEEP token in USD.
-          </Tooltip>
-          {isAPYFetching ? (
-            <Skeleton tag="h2" shining color="grey-10" />
-          ) : (
-            <h2 className={"liquidity__info-tile__title text-mint-100"}>
-              {formattedApy === Infinity ? (
-                <span>&#8734;</span>
-              ) : (
-                <CountUp
-                  end={formattedApy}
-                  // Save previously ended number to start every new animation from it.
-                  preserveValue
-                  decimals={2}
-                  duration={1}
-                  formattingFn={formattingFn}
-                />
-              )}
-            </h2>
-          )}
+          <APY.Tooltip className="liquidity__info-tile__tooltip" />
+          <APY
+            apy={apy}
+            isFetching={isAPYFetching}
+            className="liquidity__info-tile__title text-mint-100"
+          />
           <h6>Estimate of pool apy</h6>
         </div>
         <div className={"liquidity__info-tile bg-mint-10"}>
-          {isFetching ? (
-            <Skeleton tag="h2" shining color="grey-10" />
-          ) : (
-            <h2 className={"liquidity__info-tile__title text-mint-100"}>
-              <CountUp
-                end={formattedPercentageOfTotalPool}
-                // Save previously ended number to start every new animation from it.
-                preserveValue
-                decimals={2}
-                duration={1}
-                formattingFn={formattingFn}
-              />
-            </h2>
-          )}
+          <ShareOfPool
+            className="liquidity__info-tile__title text-mint-100"
+            percentageOfTotalPool={percentageOfTotalPool}
+            isFetching={isFetching}
+          />
           <h6>Your share of POOL</h6>
         </div>
       </div>
