+++ solidity/dashboard/src/components/liquidity/index.js
@@ -1,0 +1,4 @@
+import APY from "./APY"
+import ShareOfPool from "./ShareOfPool"
+
+export { APY, ShareOfPool }
